Added support for Encoder with custom differences
Added method GetEncoder2 which gets 2 encoding names.
The name of the base encoding and the new encoding name.

Added the string differences_str to the HPDF_BasicEncoderAttr_Rec struct.
If the differences_str is set, it will be used as the difference entry of the encoding.

fixes #12

Added new and alternative Unicode-Glyph-Mappings

The hpdf_encoder.c gets new entrys for the HPDF_UNICODE_GRYPH_NAME_MAP
variable.
These are partially new and partially alternative Unicode-to-Glyph
mappings.

Added HPDF_GetEncoder2 function to libhpdf.def